+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Added `send-activity` command.
+- Added `create-activity` command with an option to generate LitePub relay activities.
 
 ### Changed